LayoutTests/inspector/dom-debugger/xhr-breakpoints-expected.txt

Tests for XHR breakpoints.


== Running test suite: XHRBreakpoints
-- Running test case: BreakOnAllXMLHttpRequests
Set pause on all XMLHttpRequests.
Sending XMLHttpRequest.
PAUSED:
PASS: Pause reason should be XHR.
PASS: Pause data URL should be empty.
CALL STACK:
0: [F] loadResourceXHR
1: [P] Global Code
-- Running test teardown.

-- Running test case: BreakOnXMLHttpRequestContainingURL
PASS: Breakpoint should not be disabled initially.
PASS: Added breakpoint for URL: data
Sending XMLHttpRequest.
PAUSED:
PASS: Pause reason should be XHR.
PASS: Pause data URL should match breakpoint URL.
CALL STACK:
0: [F] loadResourceXHR
1: [P] Global Code
-- Running test teardown.

-- Running test case: ShouldNotPauseOnDisabledBreakpoint
PASS: Breakpoint should not be disabled initially.
Disabling breakpoint.
PASS: Added breakpoint for URL: data
Sending XMLHttpRequest.
Wait for evaluate in page to return.
PASS: Should not pause for disabled breakpoint.
-- Running test teardown.

-- Running test case: ShouldNotPauseOnXMLHttpRequestNotContainingURL
PASS: Breakpoint should not be disabled initially.
PASS: Added breakpoint for URL: nonexistant-url
Sending XMLHttpRequest.
Wait for evaluate in page to return.
PASS: Should not pause for breakpoint with different URL.
-- Running test teardown.

-- Running test case: RemoveBreakpoint
PASS: Breakpoint should not be disabled initially.
PASS: Added breakpoint for URL: data
Remove breakpoint.
Wait for evaluate in page to return.
PASS: Should not pause for removed breakpoint.
-- Running test teardown.
